+++ sipservice/src/main/java/net/gotev/sipservice/constants/PushMessageConstants.kt
@@ -1,4 +1,4 @@
-package net.gotev.sipservice.utility
+package net.gotev.sipservice.constants
 
 public object PushMessageConstants {
     const val GENERIC_PDC_VOIP_NOTIFICATION_CHANNEL = "GENERIC_PDC_VOIP_NOTIFICATION_CHANNEL"
